lade till validering på role

+++ routes/routes.js
@@ -42,15 +42,18 @@ async function routes(fastify, options) {
     reply.header("Access-Control-Allow-Origin", "*");
     await request.jwtVerify()
     const role = request.body.role;
-    console.log(role);
     const id = new ObjectId(request.params.id);
+
+    if (role === undefined) {
+      return { message: "fälte 'role' krävs i body" }
+    }
 
     if (role != "admin") {
       return { message: "Du har inte tillgång till denna funktion", deleted: false }
     }
     const result = await collection.deleteOne({ _id: id })
     if (result.deletedCount == 0) {
-      return { message: "Bloginlägg med id '" + request.params.id + "' finns inte", deleted: false }
+      return { message: "Bloginlägg med id '" + request.params.id + "' finns inte"}
     }
     return { deleted: true }
   })
